Fixing rebase after deletion of RecordingFragment
