+++ app/src/main/java/com/example/dashcarr/presentation/tabs/map/MapFragment.kt
@@ -44,6 +44,7 @@ import java.io.BufferedReader
 import java.io.InputStreamReader
 import java.nio.charset.Charset
 import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
 
 /**
  * Fragment representing the map screen in the application. This fragment includes the map view,
@@ -95,9 +96,7 @@ class MapFragment : BaseFragment<FragmentMapBinding>(
     private var rawAcclRecord = mutableListOf<SensorData>()
     private var filtAcclRecord = mutableListOf<SensorData>()
 
-    private var count = 0
-    private var beginTime = System.nanoTime()
-    private var rc = 0.002f
+    private var stepNumber = 0
 
     // Gyroscope
     private var rawGyroData = FloatArray(3)
@@ -187,7 +186,6 @@ class MapFragment : BaseFragment<FragmentMapBinding>(
         initListeners()
         observeViewModel()
         initMap()
-        initHud()
         requestLocationPermission()
 
         binding.apply {
@@ -279,7 +277,8 @@ class MapFragment : BaseFragment<FragmentMapBinding>(
     }
 
     private fun saveToFile(stringBuilder: StringBuilder, filtered: String, sensor: String, dateTime: LocalDateTime) {
-        context?.openFileOutput("${dateTime}_${filtered}_${sensor}.csv", Context.MODE_PRIVATE).use {
+        val formatter = DateTimeFormatter.ofPattern("dd-MM-yyyy_HH-mm-ss")
+        context?.openFileOutput("${dateTime.format(formatter)}_${filtered}_${sensor}.csv", Context.MODE_PRIVATE).use {
             it?.write(stringBuilder.toString().toByteArray())
         }
     }
@@ -289,7 +288,8 @@ class MapFragment : BaseFragment<FragmentMapBinding>(
         filtered: String,
         sensor: String,
     ) {
-        recordingJson.put("${filtered}_${sensor}", "${dateTime}_${filtered}tered_${sensor}.csv")
+        val formatter = DateTimeFormatter.ofPattern("dd-MM-yyyy_HH-mm-ss")
+        recordingJson.put("${filtered}_${sensor}", "${dateTime.format(formatter)}_${filtered}tered_${sensor}.csv")
     }
 
     private fun saveToCSV() {
@@ -298,7 +298,8 @@ class MapFragment : BaseFragment<FragmentMapBinding>(
 
         val stopDateTime = LocalDateTime.now()
 
-        recordingJson.put("name", stopDateTime)
+        val formatter = DateTimeFormatter.ofPattern("dd.MM.yyyy HH:mm")
+        recordingJson.put("name", "Recording from ${stopDateTime.format(formatter)}")
         recordingJson.put("elapsed_time", elapsedTime)
         recordingJson.put("date", stopDateTime)
 
@@ -416,9 +417,7 @@ class MapFragment : BaseFragment<FragmentMapBinding>(
         rawAcclRecord = mutableListOf<SensorData>()
         filtAcclRecord = mutableListOf<SensorData>()
 
-        count = 0
-        beginTime = System.nanoTime()
-        rc = 0.002f
+        stepNumber = 0
 
         // Gyroscope
         rawGyroData = FloatArray(3)
@@ -546,12 +545,10 @@ class MapFragment : BaseFragment<FragmentMapBinding>(
 
 
     private fun filterGyroData() {
-        val tm = System.nanoTime()
-        val dt = ((tm - beginTime) / 1000000000.0f) / count
-        val alpha = rc / (rc + dt)
+        val alpha = 0.9F
         val isStarted = true
 
-        if (count == 0) {
+        if (stepNumber == 0) {
             filtGyroPrevData[0] = (1 - alpha) * rawGyroData[0]
             filtGyroPrevData[1] = (1 - alpha) * rawGyroData[1]
             filtGyroPrevData[2] = (1 - alpha) * rawGyroData[2]
@@ -570,12 +567,10 @@ class MapFragment : BaseFragment<FragmentMapBinding>(
 
 
     private fun filterAccData() {
-        val tm = System.nanoTime()
-        val dt = ((tm - beginTime) / 1000000000.0f) / count
-        val alpha = rc / (rc + dt)
+        val alpha = 0.9F
         val isStarted = true
 
-        if (count == 0) {
+        if (stepNumber == 0) {
             filtAccPrevData[0] = (1 - alpha) * rawAccData[0]
             filtAccPrevData[1] = (1 - alpha) * rawAccData[1]
             filtAccPrevData[2] = (1 - alpha) * rawAccData[2]
@@ -589,7 +584,7 @@ class MapFragment : BaseFragment<FragmentMapBinding>(
             filtAccData[1] = filtAccPrevData[1]
             filtAccData[2] = filtAccPrevData[2]
         }
-        ++count
+        ++stepNumber
     }
 
     override fun onAccuracyChanged(sensor: Sensor?, accuracy: Int) {
@@ -645,8 +640,4 @@ class MapFragment : BaseFragment<FragmentMapBinding>(
 
     }
 
-    private fun initHud() {
-//        val inflater = requireContext().getSystemService(Context.LAYOUT_INFLATER_SERVICE) as LayoutInflater
-//        inflater.inflate(R.layout.fragment_hud, binding.hudView, true)
-    }
 }
